Add analitics service
