+++ src/client/index.js
@@ -1,2 +1,3 @@
 export * from './cookieService/cookieService';
 export * as currentUserService from './currentUserService/currentUserService';
+export * as analyticsService from './analyticsService/analyticsService';
