+++ src/universal/components/GoogleAnalytics/GoogleAnalytics.js
@@ -1,5 +1,5 @@
 import React, {Component, PropTypes} from 'react';
-import forOwn from 'lodash/forOwn';
+import {initGoogleAnalytics, sendPageview} from 'client';
 
 export default class GoogleAnalytics extends Component {
   static propTypes = {
@@ -17,7 +17,9 @@ export default class GoogleAnalytics extends Component {
       return;
     }
 
-    initGoogleAnalytics(this.props.id, this.props.set);
+    const {id, set} = this.props;
+
+    initGoogleAnalytics(id, set);
 
     this.historyListener = this.context.router.listen((location) => {
       if (!location) {
@@ -56,55 +58,4 @@ export default class GoogleAnalytics extends Component {
   render() {
     return <span />;
   }
-}
-
-function initGoogleAnalytics(id, sets = {}) {
-  if (window.ga || !id) {
-    return;
-  }
-
-  window.ga = window.ga || function(){(ga.q=ga.q||[]).push(arguments)};ga.l=+new Date; // eslint-disable-line
-
-  (function loadScript() {
-    const gads = document.createElement('script');
-    gads.async = true;
-    gads.type = 'text/javascript';
-    gads.src = '//www.google-analytics.com/analytics.js';
-
-    const head = document.getElementsByTagName('head')[0];
-    head.appendChild(gads);
-  })();
-
-  window.ga('create', id, 'auto');
-
-  forOwn(sets, (value, key) => {
-    set(key, value);
-  });
-}
-
-function command(...args) {
-  if (!window.ga) {
-    throw new Error('Google analytics is not initialized');
-  }
-
-  return window.ga.apply(window.ga, args);
-}
-
-function send(what, options) {
-  return command('send', what, options);
-}
-
-function set(key, value) {
-  return command('set', key, value);
-}
-
-function sendPageview(page, title = page, user = null) {
-  const payload = {page, title};
-
-  if (user) {
-    const userId = user.objectId;
-    set('userId', userId);
-  }
-
-  return send('pageview', payload);
 }
