+++ src/universal/containers/app/ProjectDetailPage/project-detail-forms/validators.js
@@ -3,7 +3,7 @@ import {
   createValidator,
   required,
   minLength
-} from './../../../../helpers/validation';
+} from 'universal/helpers/validation';
 
 export const addTemplateFormValidator = memoize(10)(createValidator({
   name: [required, minLength(6)],
