+++ src/universal/containers/app/ProjectListPage/projectFormValidator.js
@@ -3,7 +3,7 @@ import {
   createValidator,
   required,
   minLength
-} from './../../../helpers/validation';
+} from 'universal/helpers/validation';
 
 export const projectFormValidator = memoize(10)(createValidator({
   name: [required, minLength(6)],
