+++ src/universal/containers/app/TemplateDetailPage/template-detail-forms/validators.js
@@ -2,7 +2,7 @@ import memoize from 'lru-memoize';
 import {
   createValidator,
   required
-} from './../../../../helpers/validation';
+} from 'universal/helpers/validation';
 
 export const promoteTemplateToProductionValidator = memoize(10)(createValidator({
   commitMessage: [required]
