+++ src/universal/containers/auth/SignUpPage/SignUpPage.js
@@ -3,8 +3,8 @@ import {Link} from 'react-router';
 import {connect} from 'react-redux';
 import {initialize} from 'redux-form';
 import { push as pushState } from 'react-router-redux';
-import { Paper } from './../../../components';
-import { signUpAction } from './../../../redux/reducers/auth';
+import { Paper } from 'universal/components';
+import { signUpAction } from 'universal/redux/reducers/auth';
 import SignUpForm from './SignUpForm';
 
 @connect(
