+++ src/universal/containers/auth/SignUpPage/signUpFormValidator.js
@@ -4,7 +4,7 @@ import {
   required,
   minLength,
   email
-} from './../../../helpers/validation';
+} from 'universal/helpers/validation';
 
 export const signUpFormValidator = memoize(10)(createValidator({
   email: [required, email],
