+++ webpack/prod.config.js
@@ -46,7 +46,7 @@ module.exports = {
   progress: true,
   resolve: {
     modulesDirectories: [
-      'src/universal',
+      'src',
       'node_modules'
     ],
     extensions: ['', '.json', '.js']
